+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.3 (unreleased)
 
 - Add unnumbered option to iosxe_interface_vlan resource
+- Add option to configure lists with iosxe_restconf
 
 ## 0.1.2